Geojson parse function. Raster options allowed
