--- src/services/cluster-service.js
@@ -1,35 +1,0 @@
-import { replace } from 'utils/query';
-import { CancelToken } from 'axios';
-import { get } from 'lib/request';
-
-export const fetchData = layerModel => {
-  const { layerConfig, params, sqlParams, layerRequest } = layerModel;
-
-  const layerConfigParsed =
-    layerConfig.parse === false
-      ? layerConfig
-      : JSON.parse(replace(JSON.stringify(layerConfig), params, sqlParams));
-
-  const { body } = layerConfigParsed || {};
-  const { url } = body;
-
-  if (layerRequest) {
-    layerRequest.cancel('Operation canceled by the user.');
-  }
-
-  const layerRequestSource = CancelToken.source();
-  layerModel.set('layerRequest', layerRequestSource);
-
-  const newLayerRequest = get(url, { cancelToken: layerRequestSource.token }).then(res => {
-    if (res.status > 400) {
-      console.error(res);
-      return false;
-    }
-
-    return res.data;
-  });
-
-  return newLayerRequest;
-};
-
-export default { fetchData };
